Correct model implementation
I was using the wrong replacement probability in `replace`. Instead of
depending on the number dead, the probability that an individual is
replaced with a species 1 instead of a species 2 is (population of
species 1 from beginning of time step / total population of both
species).

Now coexistence becomes rare, as
https://esajournals.onlinelibrary.wiley.com/doi/10.1890/0012-9623-93.4.373
says it should be.